FirstPR-Pro/server/requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
google-genai
